auth_service/requirements.txt
Django
django-rest-framework
psycopg2-binary
django-oauth-toolkit
Pillow
